refactor: renamed all instances of container manager to session manager
