--- apps/container-manager/go.mod
@@ -1,3 +1,0 @@
-module container-manager
-
-go 1.23.6
