+++ apps/session-manager/.gitignore
