+++ apps/session-manager/go.mod
@@ -1,0 +1,3 @@
+module session-manager
+
+go 1.23.6
