+++ apps/session-manager/handlers/is-running/is-running.go
