+++ apps/session-manager/handlers/kill/kill.go
