+++ apps/session-manager/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"container-manager/handlers/create"
-	"container-manager/handlers/exec"
-	is_running "container-manager/handlers/is-running"
-	"container-manager/handlers/kill"
 	"fmt"
 	"net/http"
+	"session-manager/handlers/create"
+	"session-manager/handlers/exec"
+	is_running "session-manager/handlers/is-running"
+	"session-manager/handlers/kill"
 )
 
 func main() {
