+++ apps/session-manager/utils/utils.go
